Move RtcStats into separate module
