+++ crates/native/src/api/mod.rs
@@ -3,6 +3,7 @@
 pub mod media_device_info;
 pub mod media_display_info;
 pub mod rtc_ice_candidate_stats;
+pub mod rtc_stats;
 pub mod rtp_capabilities;
 pub mod rtp_codec_capability;
 pub mod rtp_header_extension_capability;
@@ -24,6 +25,11 @@ pub use self::{
     media_display_info::{MediaDisplayInfo, enumerate_displays},
     rtc_ice_candidate_stats::{
         CandidateType, IceCandidateStats, RtcIceCandidateStats,
+    },
+    rtc_stats::{
+        IceRole, RtcInboundRtpStreamMediaType, RtcMediaSourceStatsMediaType,
+        RtcOutboundRtpStreamStatsMediaType, RtcStats,
+        RtcStatsIceCandidatePairState, RtcStatsType, get_peer_stats,
     },
     rtp_capabilities::{
         RtpCapabilities, get_rtp_receiver_capabilities,
@@ -76,83 +82,6 @@ pub static RX_TIMEOUT: Duration = Duration::from_secs(5);
 /// Indicator whether application is configured to use fake media devices.
 static FAKE_MEDIA: AtomicBool = AtomicBool::new(false);
 
-/// Fields of [`RtcStatsType::RtcMediaSourceStats`] variant.
-pub enum RtcMediaSourceStatsMediaType {
-    /// Video source fields.
-    RtcVideoSourceStats {
-        /// Width (in pixels) of the last frame originating from the source.
-        /// Before a frame has been produced this attribute is missing.
-        width: Option<u32>,
-
-        /// Height (in pixels) of the last frame originating from the source.
-        /// Before a frame has been produced this attribute is missing.
-        height: Option<u32>,
-
-        /// Total number of frames originating from this source.
-        frames: Option<u32>,
-
-        /// Number of frames originating from the source, measured during the
-        /// last second. For the first second of this object's lifetime this
-        /// attribute is missing.
-        frames_per_second: Option<f64>,
-    },
-
-    /// Audio source fields.
-    RtcAudioSourceStats {
-        /// Audio level of the media source.
-        audio_level: Option<f64>,
-
-        /// Audio energy of the media source.
-        total_audio_energy: Option<f64>,
-
-        /// Audio duration of the media source.
-        total_samples_duration: Option<f64>,
-
-        /// Only exists when the [MediaStreamTrack][1] is sourced from a
-        /// microphone where echo cancellation is applied.
-        ///
-        /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
-        echo_return_loss: Option<f64>,
-
-        /// Only exists when the [MediaStreamTrack][1] is sourced from a
-        /// microphone where echo cancellation is applied.
-        ///
-        /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
-        echo_return_loss_enhancement: Option<f64>,
-    },
-}
-
-impl From<sys::RtcMediaSourceStatsMediaType> for RtcMediaSourceStatsMediaType {
-    fn from(kind: sys::RtcMediaSourceStatsMediaType) -> Self {
-        match kind {
-            sys::RtcMediaSourceStatsMediaType::RtcVideoSourceStats {
-                width,
-                height,
-                frames,
-                frames_per_second,
-            } => Self::RtcVideoSourceStats {
-                width,
-                height,
-                frames,
-                frames_per_second,
-            },
-            sys::RtcMediaSourceStatsMediaType::RtcAudioSourceStats {
-                audio_level,
-                total_audio_energy,
-                total_samples_duration,
-                echo_return_loss,
-                echo_return_loss_enhancement,
-            } => Self::RtcAudioSourceStats {
-                audio_level,
-                total_audio_energy,
-                total_samples_duration,
-                echo_return_loss,
-                echo_return_loss_enhancement,
-            },
-        }
-    }
-}
-
 /// [MediaStreamTrack.kind][1] representation.
 ///
 /// [1]: https://w3.org/TR/mediacapture-streams#dfn-kind
@@ -170,216 +99,6 @@ impl From<sys::TrackKind> for TrackKind {
         match kind {
             sys::TrackKind::Audio => Self::Audio,
             sys::TrackKind::Video => Self::Video,
-        }
-    }
-}
-
-/// Fields of [`RtcStatsType::RtcInboundRtpStreamStats`] variant.
-pub enum RtcInboundRtpStreamMediaType {
-    /// `audio` media type fields.
-    Audio {
-        /// Indicator whether the last RTP packet whose frame was delivered to
-        /// the [RTCRtpReceiver]'s [MediaStreamTrack][1] for playout contained
-        /// voice activity or not based on the presence of the V bit in the
-        /// extension header, as defined in [RFC 6464].
-        ///
-        /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#rtcrtpreceiver-interface
-        /// [RFC 6464]: https://tools.ietf.org/html/rfc6464#page-3
-        /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
-        voice_activity_flag: Option<bool>,
-
-        /// Total number of samples that have been received on this RTP stream.
-        /// This includes [concealedSamples].
-        ///
-        /// [concealedSamples]: https://tinyurl.com/s6c4qe4
-        total_samples_received: Option<u64>,
-
-        /// Total number of samples that are concealed samples.
-        ///
-        /// A concealed sample is a sample that was replaced with synthesized
-        /// samples generated locally before being played out.
-        /// Examples of samples that have to be concealed are samples from lost
-        /// packets (reported in [packetsLost]) or samples from packets that
-        /// arrive too late to be played out (reported in [packetsDiscarded]).
-        ///
-        /// [packetsLost]: https://tinyurl.com/u2gq965
-        /// [packetsDiscarded]: https://tinyurl.com/yx7qyox3
-        concealed_samples: Option<u64>,
-
-        /// Total number of concealed samples inserted that are "silent".
-        ///
-        /// Playing out silent samples results in silence or comfort noise.
-        /// This is a subset of [concealedSamples].
-        ///
-        /// [concealedSamples]: https://tinyurl.com/s6c4qe4
-        silent_concealed_samples: Option<u64>,
-
-        /// Audio level of the receiving track.
-        audio_level: Option<f64>,
-
-        /// Audio energy of the receiving track.
-        total_audio_energy: Option<f64>,
-
-        /// Audio duration of the receiving track.
-        ///
-        /// For audio durations of tracks attached locally, see
-        /// [RTCAudioSourceStats][1] instead.
-        ///
-        /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcaudiosourcestats
-        total_samples_duration: Option<f64>,
-    },
-
-    /// `video` media type fields.
-    Video {
-        /// Total number of frames correctly decoded for this RTP stream, i.e.
-        /// frames that would be displayed if no frames are dropped.
-        frames_decoded: Option<u32>,
-
-        /// Total number of key frames, such as key frames in VP8 [RFC 6386] or
-        /// IDR-frames in H.264 [RFC 6184], successfully decoded for this RTP
-        /// media stream.
-        ///
-        /// This is a subset of [framesDecoded].
-        /// [framesDecoded] - [keyFramesDecoded] gives you the number of delta
-        /// frames decoded.
-        ///
-        /// [RFC 6386]: https://w3.org/TR/webrtc-stats#bib-rfc6386
-        /// [RFC 6184]: https://w3.org/TR/webrtc-stats#bib-rfc6184
-        /// [framesDecoded]: https://tinyurl.com/srfwrwt
-        /// [keyFramesDecoded]: https://tinyurl.com/qtdmhtm
-        key_frames_decoded: Option<u32>,
-
-        /// Width of the last decoded frame.
-        ///
-        /// Before the first frame is decoded this attribute is missing.
-        frame_width: Option<u32>,
-
-        /// Height of the last decoded frame.
-        ///
-        /// Before the first frame is decoded this attribute is missing.
-        frame_height: Option<u32>,
-
-        /// Sum of the interframe delays in seconds between consecutively
-        /// decoded frames, recorded just after a frame has been decoded.
-        total_inter_frame_delay: Option<f64>,
-
-        /// Number of decoded frames in the last second.
-        frames_per_second: Option<f64>,
-
-        /// Total number of Full Intra Request (FIR) packets sent by this
-        /// receiver.
-        fir_count: Option<u32>,
-
-        /// Total number of Picture Loss Indication (PLI) packets sent by this
-        /// receiver.
-        pli_count: Option<u32>,
-
-        /// Total number of Slice Loss Indication (SLI) packets sent by this
-        /// receiver.
-        sli_count: Option<u32>,
-
-        /// Number of concealment events.
-        ///
-        /// This counter increases every time a concealed sample is synthesized
-        /// after a non-concealed sample. That is, multiple consecutive
-        /// concealed samples will increase the [concealedSamples] count
-        /// multiple times but is a single concealment event.
-        ///
-        /// [concealedSamples]: https://tinyurl.com/s6c4qe4
-        concealment_events: Option<u64>,
-
-        /// Total number of complete frames received on this RTP stream.
-        ///
-        /// This metric is incremented when the complete frame is received.
-        frames_received: Option<i32>,
-    },
-}
-
-impl From<sys::RtcInboundRtpStreamMediaType> for RtcInboundRtpStreamMediaType {
-    fn from(media_type: sys::RtcInboundRtpStreamMediaType) -> Self {
-        match media_type {
-            sys::RtcInboundRtpStreamMediaType::Audio {
-                total_samples_received,
-                concealed_samples,
-                silent_concealed_samples,
-                audio_level,
-                total_audio_energy,
-                total_samples_duration,
-            } => Self::Audio {
-                total_samples_received,
-                concealed_samples,
-                silent_concealed_samples,
-                audio_level,
-                total_audio_energy,
-                total_samples_duration,
-                voice_activity_flag: None,
-            },
-            sys::RtcInboundRtpStreamMediaType::Video {
-                frames_decoded,
-                key_frames_decoded,
-                frame_width,
-                frame_height,
-                total_inter_frame_delay,
-                frames_per_second,
-                fir_count,
-                pli_count,
-                concealment_events,
-                frames_received,
-            } => Self::Video {
-                frames_decoded,
-                key_frames_decoded,
-                frame_width,
-                frame_height,
-                total_inter_frame_delay,
-                frames_per_second,
-                fir_count,
-                pli_count,
-                concealment_events,
-                frames_received,
-                sli_count: None,
-            },
-        }
-    }
-}
-
-/// Each candidate pair in the check list has a foundation and a state.
-///
-/// The foundation is the combination of the foundations of the local and remote
-/// candidates in the pair. The state is assigned once the check list for each
-/// media stream has been computed. There are five potential values that the
-/// state can have.
-pub enum RtcStatsIceCandidatePairState {
-    /// Check for this pair hasn't been performed, and it can't yet be performed
-    /// until some other check succeeds, allowing this pair to unfreeze and move
-    /// into the [`RtcStatsIceCandidatePairState::Waiting`] state.
-    Frozen,
-
-    /// Check has not been performed for this pair, and can be performed as soon
-    /// as it is the highest-priority Waiting pair on the check list.
-    Waiting,
-
-    /// Check has been sent for this pair, but the transaction is in progress.
-    InProgress,
-
-    /// Check for this pair was already done and failed, either never producing
-    /// any response or producing an unrecoverable failure response.
-    Failed,
-
-    /// Check for this pair was already done and produced a successful result.
-    Succeeded,
-}
-
-impl From<sys::RTCStatsIceCandidatePairState>
-    for RtcStatsIceCandidatePairState
-{
-    fn from(state: sys::RTCStatsIceCandidatePairState) -> Self {
-        match state {
-            sys::RTCStatsIceCandidatePairState::kFrozen => Self::Frozen,
-            sys::RTCStatsIceCandidatePairState::kWaiting => Self::Waiting,
-            sys::RTCStatsIceCandidatePairState::kInProgress => Self::InProgress,
-            sys::RTCStatsIceCandidatePairState::kFailed => Self::Failed,
-            sys::RTCStatsIceCandidatePairState::kSucceeded => Self::Succeeded,
-            _ => unreachable!(),
         }
     }
 }
@@ -404,41 +123,6 @@ impl From<sys::Protocol> for Protocol {
         match protocol {
             sys::Protocol::Tcp => Self::Tcp,
             sys::Protocol::Udp => Self::Udp,
-        }
-    }
-}
-
-/// Variants of [ICE roles][1].
-///
-/// More info in the [RFC 5245].
-///
-/// [RFC 5245]: https://tools.ietf.org/html/rfc5245
-/// [1]: https://w3.org/TR/webrtc#dom-icetransport-role
-#[derive(Clone, Copy, Debug)]
-pub enum IceRole {
-    /// Agent whose role, as defined by [Section 3 in RFC 5245][1], has not yet
-    /// been determined.
-    ///
-    /// [1]: https://tools.ietf.org/html/rfc5245#section-3
-    Unknown,
-
-    /// Controlling agent as defined by [Section 3 in RFC 5245][1].
-    ///
-    /// [1]: https://tools.ietf.org/html/rfc5245#section-3
-    Controlling,
-
-    /// Controlled agent as defined by [Section 3 in RFC 5245][1].
-    ///
-    /// [1]: https://tools.ietf.org/html/rfc5245#section-3
-    Controlled,
-}
-
-impl From<sys::IceRole> for IceRole {
-    fn from(role: sys::IceRole) -> Self {
-        match role {
-            sys::IceRole::Unknown => Self::Unknown,
-            sys::IceRole::Controlling => Self::Controlling,
-            sys::IceRole::Controlled => Self::Controlled,
         }
     }
 }
@@ -740,557 +424,6 @@ impl From<sys::ScalabilityMode> for ScalabilityMode {
             sys::ScalabilityMode::kS3T3h => Self::S3T3h,
             _ => unreachable!(),
         }
-    }
-}
-
-/// Fields of [`RtcStatsType::RtcOutboundRtpStreamStats`] variant.
-pub enum RtcOutboundRtpStreamStatsMediaType {
-    /// `audio` media type fields.
-    Audio {
-        /// Total number of samples that have been sent over the RTP stream.
-        total_samples_sent: Option<u64>,
-
-        /// Whether the last RTP packet sent contained voice activity or not
-        /// based on the presence of the V bit in the extension header.
-        voice_activity_flag: Option<bool>,
-    },
-
-    /// `video` media type fields.
-    Video {
-        /// Width of the last encoded frame.
-        ///
-        /// The resolution of the encoded frame may be lower than the media
-        /// source (see [RTCVideoSourceStats.width][1]).
-        ///
-        /// Before the first frame is encoded this attribute is missing.
-        ///
-        /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcvideosourcestats-width
-        frame_width: Option<u32>,
-
-        /// Height of the last encoded frame.
-        ///
-        /// The resolution of the encoded frame may be lower than the media
-        /// source (see [RTCVideoSourceStats.height][1]).
-        ///
-        /// Before the first frame is encoded this attribute is missing.
-        ///
-        /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcvideosourcestats-height
-        frame_height: Option<u32>,
-
-        /// Number of encoded frames during the last second.
-        ///
-        /// This may be lower than the media source frame rate (see
-        /// [RTCVideoSourceStats.framesPerSecond][1]).
-        ///
-        /// [1]: https://tinyurl.com/rrmkrfk
-        frames_per_second: Option<f64>,
-    },
-}
-
-impl From<sys::RtcOutboundRtpStreamStatsMediaType>
-    for RtcOutboundRtpStreamStatsMediaType
-{
-    fn from(kind: sys::RtcOutboundRtpStreamStatsMediaType) -> Self {
-        use sys::RtcOutboundRtpStreamStatsMediaType as T;
-
-        match kind {
-            T::Audio => Self::Audio {
-                total_samples_sent: None,
-                voice_activity_flag: None,
-            },
-            T::Video { frame_width, frame_height, frames_per_second } => {
-                Self::Video { frame_width, frame_height, frames_per_second }
-            }
-        }
-    }
-}
-
-/// All known types of [`RtcStats`].
-///
-/// [List of all RTCStats types on W3C][1].
-///
-/// [1]: https://w3.org/TR/webrtc-stats#rtctatstype-%2A
-pub enum RtcStatsType {
-    /// Statistics for the media produced by a [MediaStreamTrack][1] that is
-    /// currently attached to an [RTCRtpSender]. This reflects the media that is
-    /// fed to the encoder after [getUserMedia()] constraints have been applied
-    /// (i.e. not the raw media produced by the camera).
-    ///
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#rtcrtpsender-interface
-    /// [getUserMedia()]: https://tinyurl.com/sngpyr6
-    /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
-    RtcMediaSourceStats {
-        /// Value of the [MediaStreamTrack][1]'s ID attribute.
-        ///
-        /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
-        track_identifier: Option<String>,
-
-        /// Fields which should be in these [`RtcStats`] based on their `kind`.
-        kind: RtcMediaSourceStatsMediaType,
-    },
-
-    /// ICE remote candidate statistics related to the [RTCIceTransport]
-    /// objects.
-    ///
-    /// A remote candidate is [deleted][1] when the [RTCIceTransport] does an
-    /// ICE restart, and the candidate is no longer a member of any non-deleted
-    /// candidate pair.
-    ///
-    /// [RTCIceTransport]: https://w3.org/TR/webrtc#dom-rtcicetransport
-    /// [1]: https://w3.org/TR/webrtc-stats#dfn-deleted
-    RtcIceCandidateStats(RtcIceCandidateStats),
-
-    /// Statistics for an outbound [RTP] stream that is currently sent with
-    /// [RTCPeerConnection] object.
-    ///
-    /// When there are multiple [RTP] streams connected to the same sender, such
-    /// as when using simulcast or RTX, there will be one
-    /// [RTCOutboundRtpStreamStats][5] per RTP stream, with distinct values of
-    /// the [SSRC] attribute, and all these senders will have a reference to the
-    /// same "sender" object (of type [RTCAudioSenderStats][1] or
-    /// [RTCVideoSenderStats][2]) and "track" object (of type
-    /// [RTCSenderAudioTrackAttachmentStats][3] or
-    /// [RTCSenderVideoTrackAttachmentStats][4]).
-    ///
-    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-    /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-    /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-    /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcaudiosenderstats
-    /// [2]: https://w3.org/TR/webrtc-stats#dom-rtcvideosenderstats
-    /// [3]: https://tinyurl.com/sefa5z4
-    /// [4]: https://tinyurl.com/rkuvpl4
-    /// [5]: https://w3.org/TR/webrtc-stats#dom-rtcoutboundrtpstreamstats
-    RtcOutboundRtpStreamStats {
-        /// ID of the stats object representing the current track attachment to
-        /// the sender of the stream.
-        track_id: Option<String>,
-
-        /// Fields which should be in these [`RtcStats`] based on their
-        /// `media_type`.
-        media_type: RtcOutboundRtpStreamStatsMediaType,
-
-        /// Total number of bytes sent for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        bytes_sent: Option<u64>,
-
-        /// Total number of RTP packets sent for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        packets_sent: Option<u32>,
-
-        /// ID of the stats object representing the track currently attached to
-        /// the sender of the stream.
-        media_source_id: Option<String>,
-    },
-
-    /// Statistics for an inbound [RTP] stream that is currently received with
-    /// [RTCPeerConnection] object.
-    ///
-    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-    /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-    RtcInboundRtpStreamStats {
-        /// ID of the stats object representing the receiving track.
-        remote_id: Option<String>,
-
-        /// Total number of bytes received for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        bytes_received: Option<u64>,
-
-        /// Total number of RTP data packets received for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        packets_received: Option<u32>,
-
-        /// Total number of RTP data packets for this [SSRC] that have been lost
-        /// since the beginning of reception.
-        ///
-        /// This number is defined to be the number of packets expected less the
-        /// number of packets actually received, where the number of packets
-        /// received includes any which are late or duplicates. Thus, packets
-        /// that arrive late are not counted as lost, and the loss
-        /// **may be negative** if there are duplicates.
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        packets_lost: Option<u64>,
-
-        /// Packet jitter measured in seconds for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        jitter: Option<f64>,
-
-        /// Total number of seconds that have been spent decoding the
-        /// [framesDecoded] frames of the stream.
-        ///
-        /// The average decode time can be calculated by dividing this value
-        /// with [framesDecoded]. The time it takes to decode one frame is the
-        /// time passed between feeding the decoder a frame and the decoder
-        /// returning decoded data for that frame.
-        ///
-        /// [framesDecoded]: https://tinyurl.com/srfwrwt
-        total_decode_time: Option<f64>,
-
-        /// Total number of audio samples or video frames that have come out of
-        /// the jitter buffer (increasing [jitterBufferDelay]).
-        ///
-        /// [jitterBufferDelay]: https://tinyurl.com/qvoojt5
-        jitter_buffer_emitted_count: Option<u64>,
-
-        /// Fields which should be in these [`RtcStats`] based on their
-        /// `media_type`.
-        media_type: Option<RtcInboundRtpStreamMediaType>,
-    },
-
-    /// ICE candidate pair statistics related to the [RTCIceTransport] objects.
-    ///
-    /// A candidate pair that is not the current pair for a transport is
-    /// [deleted] when the [RTCIceTransport] does an ICE restart, at the time
-    /// the state changes to [new].
-    ///
-    /// The candidate pair that is the current pair for a transport is [deleted]
-    /// after an ICE restart when the [RTCIceTransport] switches to using a
-    /// candidate pair generated from the new candidates; this time doesn't
-    /// correspond to any other externally observable event.
-    ///
-    /// [deleted]: https://w3.org/TR/webrtc-stats#dfn-deleted
-    /// [new]: https://w3.org/TR/webrtc#dom-rtcicetransportstate-new
-    /// [RTCIceTransport]: https://w3.org/TR/webrtc#dom-rtcicetransport
-    RtcIceCandidatePairStats {
-        /// State of the checklist for the local and remote candidates in a
-        /// pair.
-        state: RtcStatsIceCandidatePairState,
-
-        /// Related to updating the nominated flag described in
-        /// [Section 7.1.3.2.4 of RFC 5245][1].
-        ///
-        /// [1]: https://tools.ietf.org/html/rfc5245#section-7.1.3.2.4
-        nominated: Option<bool>,
-
-        /// Total number of payload bytes sent on this candidate pair, i.e. not
-        /// including headers or padding.
-        bytes_sent: Option<u64>,
-
-        /// Total number of payload bytes received on this candidate pair, i.e.
-        /// not including headers or padding.
-        bytes_received: Option<u64>,
-
-        /// Sum of all round trip time measurements in seconds since the
-        /// beginning of the session, based on STUN connectivity check
-        /// [STUN-PATH-CHAR] responses ([responsesReceived][2]), including those
-        /// that reply to requests that are sent in order to verify consent
-        /// [RFC 7675].
-        ///
-        /// The average round trip time can be computed from
-        /// [totalRoundTripTime][1] by dividing it by [responsesReceived][2].
-        ///
-        /// [STUN-PATH-CHAR]: https://w3.org/TR/webrtc-stats#bib-stun-path-char
-        /// [RFC 7675]: https://tools.ietf.org/html/rfc7675
-        /// [1]: https://tinyurl.com/tgr543a
-        /// [2]: https://tinyurl.com/r3zo2um
-        total_round_trip_time: Option<f64>,
-
-        /// Latest round trip time measured in seconds, computed from both STUN
-        /// connectivity checks [STUN-PATH-CHAR], including those that are sent
-        /// for consent verification [RFC 7675].
-        ///
-        /// [STUN-PATH-CHAR]: https://w3.org/TR/webrtc-stats#bib-stun-path-char
-        /// [RFC 7675]: https://tools.ietf.org/html/rfc7675
-        current_round_trip_time: Option<f64>,
-
-        /// Calculated by the underlying congestion control by combining the
-        /// available bitrate for all the outgoing RTP streams using this
-        /// candidate pair. The bitrate measurement does not count the size of
-        /// the IP or other transport layers like TCP or UDP. It is similar to
-        /// the TIAS defined in [RFC 3890], i.e. it is measured in bits per
-        /// second and the bitrate is calculated over a 1 second window.
-        ///
-        /// Implementations that do not calculate a sender-side estimate MUST
-        /// leave this undefined. Additionally, the value MUST be undefined for
-        /// candidate pairs that were never used. For pairs in use, the estimate
-        /// is normally no lower than the bitrate for the packets sent at
-        /// [lastPacketSentTimestamp][1], but might be higher. For candidate
-        /// pairs that are not currently in use but were used before,
-        /// implementations MUST return undefined.
-        ///
-        /// [RFC 3890]: https://tools.ietf.org/html/rfc3890
-        /// [1]: https://tinyurl.com/rfc72eh
-        available_outgoing_bitrate: Option<f64>,
-    },
-
-    /// Transport statistics related to the [RTCPeerConnection] object.
-    ///
-    /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-    RtcTransportStats {
-        /// Total number of packets sent over this transport.
-        packets_sent: Option<u64>,
-
-        /// Total number of packets received on this transport.
-        packets_received: Option<u64>,
-
-        /// Total number of payload bytes sent on this [RTCPeerConnection], i.e.
-        /// not including headers or padding.
-        ///
-        /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-        bytes_sent: Option<u64>,
-
-        /// Total number of bytes received on this [RTCPeerConnection], i.e. not
-        /// including headers or padding.
-        ///
-        /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-        bytes_received: Option<u64>,
-
-        /// Set to the current value of the [role][1] of the underlying
-        /// [RTCDtlsTransport][2]'s [transport][3].
-        ///
-        /// [1]: https://w3.org/TR/webrtc#dom-icetransport-role
-        /// [2]: https://w3.org/TR/webrtc#rtcdtlstransport-interface
-        /// [3]: https://w3.org/TR/webrtc#dom-rtcdtlstransport-icetransport
-        ice_role: Option<IceRole>,
-    },
-
-    /// Statistics for the remote endpoint's inbound [RTP] stream corresponding
-    /// to an outbound stream that is currently sent with [RTCPeerConnection]
-    /// object.
-    ///
-    /// It is measured at the remote endpoint and reported in a RTCP Receiver
-    /// Report (RR) or RTCP Extended Report (XR).
-    ///
-    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-    /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-    RtcRemoteInboundRtpStreamStats {
-        /// [localId] is used for looking up the local
-        /// [RTCOutboundRtpStreamStats][1] object for the same [SSRC].
-        ///
-        /// [localId]: https://tinyurl.com/r8uhbo9
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcoutboundrtpstreamstats
-        local_id: Option<String>,
-
-        /// Packet jitter measured in seconds for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        jitter: Option<f64>,
-
-        /// Estimated round trip time for this [SSRC] based on the RTCP
-        /// timestamps in the RTCP Receiver Report (RR) and measured in seconds.
-        /// Calculated as defined in [Section 6.4.1 of RFC 3550][1].
-        /// If no RTCP Receiver Report is received with a DLSR value other than
-        /// 0, the round trip time is left undefined.
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        /// [1]: https://tools.ietf.org/html/rfc3550#section-6.4.1
-        round_trip_time: Option<f64>,
-
-        /// Fraction packet loss reported for this [SSRC].
-        /// Calculated as defined in [Section 6.4.1 of RFC 3550][1] and
-        /// [Appendix A.3][2].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        /// [1]: https://tools.ietf.org/html/rfc3550#section-6.4.1
-        /// [2]: https://tools.ietf.org/html/rfc3550#appendix-A.3
-        fraction_lost: Option<f64>,
-
-        /// Total number of RTCP RR blocks received for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        reports_received: Option<u64>,
-
-        /// Total number of RTCP RR blocks received for this [SSRC] that contain
-        /// a valid round trip time. This counter will increment if the
-        /// [roundTripTime] is undefined.
-        ///
-        /// [roundTripTime]: https://tinyurl.com/ssg83hq
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        round_trip_time_measurements: Option<i32>,
-    },
-
-    /// Statistics for the remote endpoint's outbound [RTP] stream corresponding
-    /// to an inbound stream that is currently received with [RTCPeerConnection]
-    /// object.
-    ///
-    /// It is measured at the remote endpoint and reported in an RTCP Sender
-    /// Report (SR).
-    ///
-    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-    /// [RTCPeerConnection]: https://w3.org/TR/webrtc#dom-rtcpeerconnection
-    RtcRemoteOutboundRtpStreamStats {
-        /// [localId] is used for looking up the local
-        /// [RTCInboundRtpStreamStats][1] object for the same [SSRC].
-        ///
-        /// [localId]: https://tinyurl.com/vu9tb2e
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        /// [1]: https://w3.org/TR/webrtc-stats#dom-rtcinboundrtpstreamstats
-        local_id: Option<String>,
-
-        /// [remoteTimestamp] (as [HIGHRES-TIME]) is the remote timestamp at
-        /// which these statistics were sent by the remote endpoint. This
-        /// differs from timestamp, which represents the time at which the
-        /// statistics were generated or received by the local endpoint. The
-        /// [remoteTimestamp], if present, is derived from the NTP timestamp in
-        /// an RTCP Sender Report (SR) block, which reflects the remote
-        /// endpoint's clock. That clock may not be synchronized with the local
-        /// clock.
-        ///
-        /// [HIGRES-TIME]: https://w3.org/TR/webrtc-stats#bib-highres-time
-        /// [remoteTimestamp]: https://tinyurl.com/rzlhs87
-        remote_timestamp: Option<f64>,
-
-        /// Total number of RTCP SR blocks sent for this [SSRC].
-        ///
-        /// [SSRC]: https://w3.org/TR/webrtc-stats#dfn-ssrc
-        reports_sent: Option<u64>,
-    },
-
-    /// Unimplemented stats.
-    Unimplemented,
-}
-
-impl From<sys::RtcStatsType> for RtcStatsType {
-    #[expect(clippy::too_many_lines, reason = "trivial code")]
-    fn from(kind: sys::RtcStatsType) -> Self {
-        use sys::RtcStatsType as T;
-
-        match kind {
-            T::RtcMediaSourceStats { track_identifier, kind } => {
-                Self::RtcMediaSourceStats {
-                    track_identifier,
-                    kind: kind.into(),
-                }
-            }
-            T::RtcIceCandidateStats(stats) => match stats {
-                sys::RtcIceCandidateStats::RtcLocalIceCandidateStats(
-                    candidate,
-                ) => Self::RtcIceCandidateStats(RtcIceCandidateStats::Local(
-                    candidate.into(),
-                )),
-                sys::RtcIceCandidateStats::RtcRemoteIceCandidateStats(
-                    candidate,
-                ) => Self::RtcIceCandidateStats(RtcIceCandidateStats::Remote(
-                    candidate.into(),
-                )),
-            },
-            T::RtcOutboundRtpStreamStats {
-                track_id,
-                media_type,
-                bytes_sent,
-                packets_sent,
-                media_source_id,
-            } => Self::RtcOutboundRtpStreamStats {
-                track_id,
-                media_type: media_type.into(),
-                bytes_sent,
-                packets_sent,
-                media_source_id,
-            },
-            T::RtcInboundRtpStreamStats {
-                remote_id,
-                bytes_received,
-                packets_received,
-                total_decode_time,
-                jitter_buffer_emitted_count,
-                media_type,
-            } => Self::RtcInboundRtpStreamStats {
-                remote_id,
-                bytes_received,
-                packets_received,
-                total_decode_time,
-                jitter_buffer_emitted_count,
-                media_type: media_type.map(RtcInboundRtpStreamMediaType::from),
-                packets_lost: None,
-                jitter: None,
-            },
-            T::RtcIceCandidatePairStats {
-                state,
-                nominated,
-                bytes_sent,
-                bytes_received,
-                total_round_trip_time,
-                current_round_trip_time,
-                available_outgoing_bitrate,
-            } => Self::RtcIceCandidatePairStats {
-                state: state.into(),
-                nominated,
-                bytes_sent,
-                bytes_received,
-                total_round_trip_time,
-                current_round_trip_time,
-                available_outgoing_bitrate,
-            },
-            T::RtcTransportStats {
-                packets_sent,
-                packets_received,
-                bytes_sent,
-                bytes_received,
-            } => Self::RtcTransportStats {
-                packets_sent,
-                packets_received,
-                bytes_sent,
-                bytes_received,
-                ice_role: None,
-            },
-            T::RtcRemoteInboundRtpStreamStats {
-                local_id,
-                round_trip_time,
-                fraction_lost,
-                round_trip_time_measurements,
-            } => Self::RtcRemoteInboundRtpStreamStats {
-                local_id,
-                round_trip_time,
-                fraction_lost,
-                round_trip_time_measurements,
-                jitter: None,
-                reports_received: None,
-            },
-            T::RtcRemoteOutboundRtpStreamStats {
-                local_id,
-                remote_timestamp,
-                reports_sent,
-            } => Self::RtcRemoteOutboundRtpStreamStats {
-                local_id,
-                remote_timestamp,
-                reports_sent,
-            },
-            T::Unimplemented => Self::Unimplemented,
-        }
-    }
-}
-
-/// Represents the [stats object] constructed by inspecting a specific
-/// [monitored object].
-///
-/// [Full doc on W3C][1].
-///
-/// [stats object]: https://w3.org/TR/webrtc-stats#dfn-stats-object
-/// [monitored object]: https://w3.org/TR/webrtc-stats#dfn-monitored-object
-/// [1]: https://w3.org/TR/webrtc#rtcstats-dictionary
-pub struct RtcStats {
-    /// Unique ID that is associated with the object that was inspected to
-    /// produce this [RTCStats] object.
-    ///
-    /// [RTCStats]: https://w3.org/TR/webrtc#dom-rtcstats
-    pub id: String,
-
-    /// Timestamp associated with this object.
-    ///
-    /// The time is relative to the UNIX epoch (Jan 1, 1970, UTC).
-    ///
-    /// For statistics that came from a remote source (e.g., from received RTCP
-    /// packets), timestamp represents the time at which the information
-    /// arrived at the local endpoint. The remote timestamp can be found in an
-    /// additional field in an [`RtcStats`]-derived dictionary, if applicable.
-    pub timestamp_us: i64,
-
-    /// Actual stats of these [`RtcStats`].
-    ///
-    /// All possible stats are described in the [`RtcStatsType`] enum.
-    pub kind: RtcStatsType,
-}
-
-impl From<sys::RtcStats> for RtcStats {
-    fn from(stats: sys::RtcStats) -> Self {
-        let sys::RtcStats { id, timestamp_us, kind } = stats;
-        Self { id, timestamp_us, kind: RtcStatsType::from(kind) }
     }
 }
 
@@ -2448,19 +1581,6 @@ pub fn get_transceiver_direction(
     transceiver: RustOpaque<Arc<RtpTransceiver>>,
 ) -> RtpTransceiverDirection {
     transceiver.direction().into()
-}
-
-/// Returns [`RtcStats`] of the [`PeerConnection`] by its ID.
-#[expect(clippy::needless_pass_by_value, reason = "FFI")]
-pub fn get_peer_stats(
-    peer: RustOpaque<Arc<PeerConnection>>,
-) -> anyhow::Result<Vec<RtcStats>> {
-    let (tx, rx) = mpsc::channel();
-
-    peer.get_stats(tx);
-    let report = rx.recv_timeout(RX_TIMEOUT)?;
-
-    Ok(report.get_stats()?.into_iter().map(RtcStats::from).collect())
 }
 
 /// Irreversibly marks the specified [`RtcRtpTransceiver`] as stopping, unless
